requirements.txt
# חבילות בסיסיות
python-dotenv==1.0.0
python-telegram-bot==21.0.1
openai==1.14.0
anthropic==0.18.1
logfire==0.16.0
logfire[httpx]==0.16.0

# חבילות מסד נתונים
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
asyncpg==0.30.0
pgvector==0.3.6

# חבילות עזר
tabulate==0.9.0
numpy==2.2.3
pydantic==2.6.1

# חבילות לטיפול בסוגי קבצים שונים
pypdf==4.0.1
python-docx==1.1.0
python-pptx==1.0.1
openpyxl==3.1.2
beautifulsoup4==4.12.2
html2text==2024.2.26

# חבילות ווקומרס
WooCommerce==3.0.0
